requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-pptx
selenium
webdriver-manager
